added geocoder and showing maps

+++ db/migrate/20151209140838_add_latitude_to_store.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToStore < ActiveRecord::Migration
   def change
-    add_column :stores, :latitude, :integer
+    add_column :stores, :latitude, :float
   end
 end
